finalizing category

+++ kotlin-backend-developer-2023/src/main/kotlin/com/tqi/kotlinbackenddeveloper2023/domain/service/impl/CategoryService.kt
@@ -1,10 +1,10 @@
 package com.tqi.kotlinbackenddeveloper2023.domain.service.impl
 
+import com.tqi.kotlinbackenddeveloper2023.domain.exceptions.BusinessException
 import com.tqi.kotlinbackenddeveloper2023.domain.model.Category
 import com.tqi.kotlinbackenddeveloper2023.domain.repository.CategoryRepository
 import com.tqi.kotlinbackenddeveloper2023.domain.service.ICategoryService
 import org.springframework.stereotype.Service
-import java.util.*
 
 @Service
 class CategoryService(private val categoryRepository: CategoryRepository) : ICategoryService {
@@ -14,7 +14,9 @@ class CategoryService(private val categoryRepository: CategoryRepository) : ICat
     }
 
     override fun alteration(category: Category): Category {
-        val existingCategory = categoryRepository.findById(category.id).get()
+        val existingCategory = categoryRepository.findById(category.id).orElseThrow {
+            throw BusinessException("id ${category.id} not found")
+        }
         existingCategory.name = category.name
         return categoryRepository.save(existingCategory)
     }
@@ -28,6 +30,11 @@ class CategoryService(private val categoryRepository: CategoryRepository) : ICat
     }
 
     override fun deleteByName(name: String) {
-        return categoryRepository.deleteByName(name)
+        if (categoryRepository.existsByName(name)) {
+            categoryRepository.deleteByName(name)
+        } else {
+            throw BusinessException("name: $name not found")
+        }
+
     }
 }
